unit tests for AddLesson component 95%
